changed NotificationSettingsAction to use sealed class for configuration
