+++ notifications/src/main/java/dk/shape/games/notifications/actions/NotificationSettingsAction.kt
@@ -3,24 +3,10 @@ package dk.shape.games.notifications.actions
 import dk.shape.games.sportsbook.offerings.common.action.Action
 import kotlinx.android.parcel.Parcelize
 
-sealed class NotificationSettingsAction : Action {
-
-    @Parcelize
-    data class IncludeAllEvents(
-        val eventIds: List<String>
-    ) : NotificationSettingsAction()
-
-    @Deprecated("This will not be used with a release of new BetHistory")
-    @Parcelize
-    object FilterBetEvents : NotificationSettingsAction()
-
-    @Parcelize
-    data class FilterBetSlip(
-        val betSlipComponentUUID: String? = null
-    ) : NotificationSettingsAction()
-
-    @Parcelize
-    object Default: NotificationSettingsAction()
+@Parcelize
+data class NotificationSettingsAction(
+    val config: NotificationSettingsActionConfig = NotificationSettingsActionConfig.Default
+) : Action {
 
     override fun isModal() = true
